CP-23309: Don't hang if xenopsd goes away
Draining the messages that may have been received before sending a new
one is wrong for a few reasons:

1) If xenopsd has closed the connection, emu-manager will hang in an
endless loop.
2) There is still a possibility of receiving a new message on the
connection after draining the messages but before sending the message
which defeats the purpose of draining messages.
3) The protocol is such that emu-manager should never be expecting a
response to a message at the same time that xenopsd could send a message
asynchronously.

Therefore, remove the code to drain messages.