CP-23309: Cleanup struct emu
Use a C99 initializer so that only some fields need to be explicitly
initialized, long lines are avoided, and it is clearer what is being
initialized.  Calculate the number of emus rather than hardcoding it.
Remove the live_check field since it is unused. Initialize the stream
field correctly -- it is an fd so it should be initialized to -1.