CA-366017: Fix broken logic, which would xen-set-global-dirty-log being sent to QEMU, of not for another bug.
STAGE_LIVE stage flag in emu-manager handled both if the emu should track dirty pages,
as well as indicating that the emu was actually migrating its content in a live fashion.
Since QEMU is not doing a live state transfer itself, it was also not being asked to
dirty page track - which it should have.
To resolve this, I have split the stage into two, with QEMU only using the first.
This inishally seemed overkill, but now think it will make the code more understanable.

This bug was masked logical AND being used in place of bitwise and in the stage test.
Also fixed some unrelated return values.